+++ nextgisweb/raster_layer/model.py
@@ -118,8 +118,6 @@ class RasterLayer(Base, Resource, SpatialLayerMixin):
         if reproject:
             cmd = ['gdalwarp', '-of', 'GTiff',
                    '-t_srs', 'EPSG:%d' % self.srs.id]
-            if ds.RasterCount == 3:
-                cmd.append('-dstalpha')
         else:
             cmd = ['gdal_translate', '-of', 'GTiff']