+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 -----
 
 - Add support of 1.1.0 version in WFS server implementation.
+- Improved handling of NODATA values in raster layer and raster style.
 - Fix infinite wait of database lock, including during vector layer deletion.
 
 3.6.0